remove fail-fast, begin writing test for view recipes
