+++ .rspec
@@ -1,3 +1,2 @@
 --color
 --require spec_helper
---fail-fast
