+++ spec/support/factories.rb
@@ -6,4 +6,6 @@ FactoryGirl.define do
     password 'password'
     password_confirmation 'password'
   end
+
+
 end
